Add jobcall
jobcall allows to synchronously run a job in the current thread by
using afb_sched_call.

The difference between jobcall and jobenter is that a jobcall ends
when its callback returns, while a jobenter ends when jobleave is
called to undo the lock.

TLDR: with jobenter one controls when the job ends.