Get rid of sub-interpreters (#6307)
It turns out sub-interpreters are not needed for our use case and adds
surprising behaviors. We basically want to be able to register callbacks to the
afb framework that are Python callables without blocking or be messed
up with threads created outside of Python.

Main changes here are:
- make sure afb functions that are prone to call Python functions in
other threads are called between Py_BEGIN_ALLOW_THREADS and
Py_END_ALLOW_THREADS (this will basically release the GIL before the
call, and restore it after)
- in all callbacks that may have been called from a thread created
outside of Python, protect Python API calls by PyGILState_Ensure() /
PyGILState_Release()

We then now have only one main Python interpreter running and the GIL
should be correctly released and acquired when needed so that Python
threads are executed and switched correctly.

Of course we could not benefit from a full parallelism between
different Python callbacks, but this is not required.